+++ src/test/resources/org/checkstyle/autofix/recipe/useenhancedswitch/mutationtests6/DiffMutationTests6.diff
@@ -1,6 +1,6 @@
 --- src/test/resources/org/checkstyle/autofix/recipe/useenhancedswitch/mutationtests6/InputMutationTests6.java
 +++ src/test/resources/org/checkstyle/autofix/recipe/useenhancedswitch/mutationtests6/OutputMutationTests6.java
-@@ -8,253 +8,174 @@
+@@ -8,96 +8,70 @@
  
  package org.checkstyle.autofix.recipe.useenhancedswitch.mutationtests6;
  
@@ -127,200 +127,5 @@
 +            case 1 -> 10;
 +            default -> throw new IllegalArgumentException("nope");
 +        };
-     }
- 
-     void multipleVarDecls(int x) {
-         int a;
-         int b;
--        // violation below, 'Switch can be replaced with enhanced switch'
-         switch (x) {
--            case 1:
--                a = 1;
--                break;
--            default:
--                a = 0;
-+            case 1 -> a = 1;
-+            default -> a = 0;
-         }
--        // violation below, 'Switch can be replaced with enhanced switch'
-         switch (x) {
--            case 1:
--                b = 1;
--                break;
--            default:
--                b = 0;
-+            case 1 -> b = 1;
-+            default -> b = 0;
-         }
-         System.out.println(a + b);
-     }
- 
-     int switchExprTraditional(int x) {
--        // violation below, 'Switch can be replaced with enhanced switch'
-         return switch (x) {
--            case 1: yield 10;
--            case 2: yield 20;
--            default: yield 0;
-+            case 1 -> 10;
-+            case 2 -> 20;
-+            default -> 0;
-         };
-     }
- 
-     void noAssignmentConversion(int x) {
-         String s = "initial";
--        // violation below, 'Switch can be replaced with enhanced switch'
-         switch (x) {
--            case 1:
--                s = "one";
--                break;
--            default:
--                s = "default";
-+            case 1 -> s = "one";
-+            default -> s = "default";
-         }
-         System.out.println(s);
-     }
- 
-     void enumAssignment(Season s) {
--        int val;
--        // violation below, 'Switch can be replaced with enhanced switch'
--        switch (s) {
--            case SPRING:
--                val = 1;
--                break;
--            case SUMMER:
--                val = 2;
--                break;
--            case FALL:
--                val = 3;
--                break;
--            case WINTER:
--                val = 4;
--                break;
--            default:
--                val = 0;
--                break;
--        }
-+        int val = switch (s) {
-+            case SPRING -> 1;
-+            case SUMMER -> 2;
-+            case FALL -> 3;
-+            case WINTER -> 4;
-+            default -> 0;
-+        };
-         System.out.println(val);
-     }
- 
-     void singleBlockCase(int x) {
--        // violation below, 'Switch can be replaced with enhanced switch'
-         switch (x) {
--            case 1: {
--                System.out.println("block");
--                break;
--            }
--            default:
--                System.out.println("default");
-+            case 1 -> System.out.println("block");
-+            default -> System.out.println("default");
-         }
-     }
- 
-     void multipleVarInOneDecl(int x) {
-         int a, b;
--        // violation below, 'Switch can be replaced with enhanced switch'
-         switch (x) {
--            case 1:
--                a = 1;
--                break;
--            default:
--                a = 0;
-+            case 1 -> a = 1;
-+            default -> a = 0;
-         }
-         System.out.println(a);
-     }
- 
-     void exhaustiveNoDefaultAssignment(Season s) {
-         int val = 0;
--        // violation below, 'Switch can be replaced with enhanced switch'
-         switch (s) {
--            case SPRING: val = 1; break;
--            case SUMMER: val = 2; break;
--            case FALL: val = 3; break;
--            case WINTER: val = 4; break;
-+            case SPRING -> val = 1;
-+            case SUMMER -> val = 2;
-+            case FALL -> val = 3;
-+            case WINTER -> val = 4;
-         }
-         System.out.println(val);
-     }
- 
-     void assignmentWithBlock(int x) {
--        int val;
--        // violation below, 'Switch can be replaced with enhanced switch'
--        switch (x) {
--            case 1: {
-+        int val = switch (x) {
-+            case 1 -> {
-                 System.out.println("block");
--                val = 1;
--                break;
-+                yield 1;
-             }
--            default:
--                val = 0;
--        }
-+            default -> 0;
-+        };
-         System.out.println(val);
-     }
- 
-     void assignmentWithSingleStatementBlock(int x) {
--        int val;
--        // violation below, 'Switch can be replaced with enhanced switch'
--        switch (x) {
--            case 1: {
--                val = 1;
--                break;
--            }
--            default:
--                val = 0;
--        }
-+        int val = switch (x) {
-+            case 1 -> 1;
-+            default -> 0;
-+        };
-         System.out.println(val);
-     }
- 
-     enum EmptyEnum {}
- 
-     void emptyEnumExhaustive(EmptyEnum e) {
--        // violation below, 'Switch can be replaced with enhanced switch'
-         switch (e) {
--            default:
--                System.out.println("default");
-+            default -> System.out.println("default");
-         }
-     }
- 
-     void mixedLabelsAssignment(Season s) {
--        int val;
--        // violation below, 'Switch can be replaced with enhanced switch'
--        switch (s) {
--            case SPRING:
--            case SUMMER:
--                val = 1;
--                break;
--            default:
--                val = 0;
--        }
-+        int val = switch (s) {
-+            case SPRING, SUMMER -> 1;
-+            default -> 0;
-+        };
-         System.out.println(val);
      }
  }
